watch: re-split check.go under the 800-line cap after taking main's #508 wholesale
check.go was 851 lines after the #508 merge (112b924) took main's version
entirely. Extracted the webhook-specific SSRF-hardening code into
webhook.go: newSafeWebhookClient, webhookCheckRedirect,
specialUseWebhookBlocks, mustParseWebhookCIDRs, isPublicWebhookIP,
webhookHTTPClient, SetWebhookHTTPClientForTest, webhookPayload, fireWebhook.
check.go keeps the general check-loop logic (CheckAll*, checkOne*,
CheckResult, PriceChecker/RoomChecker interfaces).

check.go: 595 lines. webhook.go: 263 lines. Both under 800.

go build/vet/gofmt clean. go test ./internal/watch/... and ./mcp/...: pass.
Full go test ./...: 2 unrelated pre-existing failures (internal/providers
flaky network fixture, internal/safeexec -- darwin sandbox process-
containment fixture that self-documents as unreliable on this platform),
neither touches watch/mcp.
